requirements.txt
aiofiles==25.1.0
aiogram==3.24.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
cachetools==6.2.4
certifi==2025.11.12
charset-normalizer==3.4.4
dotenv==0.9.9
frozenlist==1.8.0
google-api-core==2.29.0
google-api-python-client==2.187.0
google-auth==2.41.1
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.3
googleapis-common-protos==1.72.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.11
magic-filter==1.0.12
multidict==6.7.0
oauthlib==3.3.1
propcache==0.4.1
proto-plus==1.27.0
protobuf==6.33.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.1
pyTelegramBotAPI==4.29.1
python-dotenv==1.2.1
python-telegram-bot==22.5
pytz==2025.2
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
setuptools==80.9.0
telebot==0.0.5
telegram==0.0.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.2
yarl==1.22.0
